backend/requirements.txt
openai
PyGithub
GitPython
exa-py
python-dotenv